pyproject.toml
[tool.poetry]
name = "mirea_ninja_payment_bot"
version = "0.1.0"
description = "Bot for MIREA Ninja payment system"
authors = ["Sergey Dmitriev <51058739+0niel@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.37"
envparse = "^0.2.0"
pydantic = "^1.9.1"
requests = "^2.28.0"
python-telegram-bot = {extras = ["job-queue"], version = "^21.0.1"}
psycopg2-binary = "^2.9.3"
yappa = "^0.4.28"
pydiscourse = "^1.6.1"


[tool.poetry.dev-dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8 = "^7.0.0"
pytest = "^8.1.1"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
